+++ server/src/embedded_migrations.rs
@@ -169,4 +169,5 @@ pub const EMBEDDED_MIGRATIONS: &[(&str, &str)] = &[
     ("166_repair_txn_624598_override_evidence.sql", include_str!("../../migrations/166_repair_txn_624598_override_evidence.sql")),
     ("167_repair_recovered_helcim_payment_dates.sql", include_str!("../../migrations/167_repair_recovered_helcim_payment_dates.sql")),
     ("168_repair_txn_566034_amended_line_price.sql", include_str!("../../migrations/168_repair_txn_566034_amended_line_price.sql")),
+    ("169_restore_txn_566034_booking_reportability.sql", include_str!("../../migrations/169_restore_txn_566034_booking_reportability.sql")),
 ];